+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+`Version 0.2.3: Crashfile <https://github.com/shnizzedy/cpac-python-package/releases/tag/v0.2.3>`_
+==================================================================================================
+* ✨ Added ``group`` and ``crash`` commands
+* 🚑 Fixed a bug where pass-through flags were being mangled
+* 🖇️ Binds any directories necessary to access any paths found in pass-through CLI arguments
+
 `Version 0.2.2: Autobind <https://github.com/shnizzedy/cpac-python-package/releases/tag/v0.2.2>`_
 =================================================================================================